refactor: move relative calculation to controller
